Split omnikassa_payment.rb into two files with only one class each
